Make v1,v2,v3 internal classes and methods private
This temporarily breaks JSON serialization of channel classes (used in Phoenix and our test JVM test, for display/logging purposes)
